Separate drawing-related code
